scaliest, use scalatest over simplespec
 - simplespec is no longer maintained -
https://github.com/SimpleFinance/simplespec
